phase-2/app/pyproject.toml
[project]
name = "elastic-3am-alert-killer"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "uvicorn",
    "a2a-sdk[http-server]",
    "llama-index-core",
    "llama-index-llms-openai-like"
]

[project.optional-dependencies]
dev = [
    "llama-index-utils-workflow"
]
